Tweaked shooting + changed projectiles
+ shooting can now be held down
+ Added updated projectiles
+ Changed Aphrodite's texture
+ Changed Character Select Screen

+++ John Lemon's Haunted Jaunt/Assets/Scripts/PlayerHealth.cs
@@ -34,7 +34,7 @@ public class PlayerHealth : MonoBehaviourPunCallbacks, IPunObservable
         }
         else
         {
-            health = (int)stream.ReceiveNext();    
+            health = (float)stream.ReceiveNext();    
         }
     }
